Desktop/vote/university_evoting/requirements.txt
Django>=4.2,<5.0
python-decouple
psycopg2-binary
django-crispy-forms
django-environ
django-widget-tweaks
djangorestframework
cryptography
pyotp
fido2
locust
boto3
moto[kms]
Pillow==10.1.0
qrcode==7.4.2
reportlab==4.0.0
celery>=5.3.0
redis>=4.5.0
django-storages[boto3]>=1.14.0
